🔨 Questions 1 and 3 done
Note: for future projects like this, it might be wise to split the questions into individual .py files, considering one question will most likely share variables and functions within its subquestions.  It is probably also a good idea to write it with modularity in mind, and matrices in order to make the code reusable, and to spend less time refactoring individual lists over and over, etc.